feat: compute run preview before storing
